+++ CHANGELOG.md
@@ -8,6 +8,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - Added caching system and new table to store data and speed up repeated fetches
 
+## 1.0.6 - 2021-11-18
+### Added
+- Added support for Profile Photos.
+
+### Fixed
+- Fixed bug that occurred on prefixed SEOmatic fields.
+- SEOmatic and Profile Photo check now only occur when element is an Asset. -> Performance
+
+## 1.0.5 - 2021-11-13
+### Fixed
+- Fixed bug that occurred on Craft CMS installations with table prefixes. Thank you @gbowne-quickbase! #1
+
 ## 1.0.4 - 2021-10-25
 ### Added
 - Added support for SEOmatic.